Day 23: Add auth logic to prevent unauthorised employers to edit any job.

+++ app/Http/Controllers/JobController.php
@@ -3,11 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
-    public function index()
+     public function index()
     {
         $jobs = Job::with('employer')->latest()->simplePaginate(3);
 
@@ -50,7 +53,7 @@ class JobController extends Controller
 
     public function update(Job $job)
     {
-            // authorize (On hold...)
+        Gate::authorize('edit-job', $job);
 
     request()->validate([
         'title' => ['required', 'min:3'],
@@ -67,7 +70,7 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // authorize (On hold...)
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
